requirements.txt
langgraph==0.3.0
google-genai==1.5.0
psycopg[binary]==3.1.0
psycopg-pool==3.2.0
pydantic==2.10.0
fastapi==0.115.0
uvicorn==0.34.0
python-dotenv==1.0.0